requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
python-dotenv==0.19.2
requests==2.26.0
soupsieve==2.3.1
urllib3==1.26.7
Werkzeug==2.0.2
